Update acronym to namespace and data provider tests
