+++ exercises/acronym/src/acronym.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Exercism\Acronym;
+
+/**
+ * Abbreviates a phrase.
+ *
+ * @param string $phrase
+ * @return string
+ */
+function acronym($phrase)
+{
+    // Your code goes here
+}
